Create new class with createObservable method
